Refactored code and restructured files for cleanup, passing all tests
